refactor dev loss sanity check to separate script, remove NLLP utils, update first pass evaluate-model script
